+++ ICD.Connect.Panels.CrestronPro/Controls/TouchScreens/AbstractThreeSeriesTouchScreenControl.cs
@@ -1,0 +1,20 @@
+﻿using ICD.Connect.Devices;
+using ICD.Connect.Panels.Controls;
+using ICD.Connect.Panels.Crestron.Controls.TouchScreens;
+
+namespace ICD.Connect.Panels.CrestronPro.Controls.TouchScreens
+{
+	public abstract class AbstractThreeSeriesTouchScreenControl<TParent> : AbstractPanelControl<TParent>, IThreeSeriesTouchScreenControl
+		where TParent : IDeviceBase
+	{
+		/// <summary>
+		/// Constructor.
+		/// </summary>
+		/// <param name="parent"></param>
+		/// <param name="id"></param>
+		protected AbstractThreeSeriesTouchScreenControl(TParent parent, int id)
+			: base(parent, id)
+		{
+		}
+	}
+}