+++ ICD.Connect.Panels.CrestronPro/TriListAdapters/X60/Tsw760/Tsw760Adapter.cs
@@ -6,7 +6,7 @@ namespace ICD.Connect.Panels.CrestronPro.TriListAdapters.X60.Tsw760
 {
 #if SIMPLSHARP
 	public sealed class Tsw760Adapter :
-		AbstractTswX60BaseClassAdapter<Crestron.SimplSharpPro.UI.Tsw760, Tsw760AdapterSettings>
+		AbstractTswX60BaseClassAdapter<global::Crestron.SimplSharpPro.UI.Tsw760, Tsw760AdapterSettings>
 	{
 		/// <summary>
 		/// Creates an instance of the wrapped trilist.
@@ -14,9 +14,9 @@ namespace ICD.Connect.Panels.CrestronPro.TriListAdapters.X60.Tsw760
 		/// <param name="ipid"></param>
 		/// <param name="controlSystem"></param>
 		/// <returns></returns>
-		protected override Crestron.SimplSharpPro.UI.Tsw760 InstantiateTriList(byte ipid, CrestronControlSystem controlSystem)
+		protected override global::Crestron.SimplSharpPro.UI.Tsw760 InstantiateTriList(byte ipid, CrestronControlSystem controlSystem)
 		{
-			return new Crestron.SimplSharpPro.UI.Tsw760(ipid, controlSystem);
+			return new global::Crestron.SimplSharpPro.UI.Tsw760(ipid, controlSystem);
 		}
 	}
 #else