+++ ICD.Connect.Panels/Controls/ISigControl.cs
@@ -1,0 +1,8 @@
+﻿using ICD.Connect.Devices.Controls;
+
+namespace ICD.Connect.Panels.Controls
+{
+	public interface ISigControl : IDeviceControl, ISigInputOutput
+	{
+	}
+}